Change to int values for QImage and setDotsPerMeter

+++ nezzle/io/image.py
@@ -27,14 +27,14 @@ def write_image(net,
                        image_height,
                        QImage.Format_ARGB32_Premultiplied)
     else:
-        image = QImage((scale_width/100.0) * brect.width(),
-                       (scale_height/100.0) * brect.height(),
+        image = QImage(int((scale_width/100.0) * brect.width()),
+                       int((scale_height/100.0) * brect.height()),
                        QImage.Format_ARGB32_Premultiplied)
 
     # [REF] http://stackoverflow.com/a/13425280/4136588
     # dpm = 300 / 0.0254 # ~300 DPI
-    dpm_width = dpi_width / 0.0254
-    dpm_height = dpi_height / 0.0254
+    dpm_width = int(dpi_width / 0.0254)
+    dpm_height = int(dpi_height / 0.0254)
     image.setDotsPerMeterX(dpm_width)
     image.setDotsPerMeterY(dpm_height)
